Make an API change + make a different projection for spring. (#12)
* Make an API change + make a different projection for spring.

From now, constraints will return **all** their points and projected
points, since we're always working with a vertex.

Also, for mass spring with equal masses or both ends the projection in
the middle is more stable for the constraint. We assume the masses are
equal for now. If it is not, it gets a bit unstable.

* Automated autopep8 fixes

Co-authored-by: autopep8-bot <autopep8-bot@users.noreply.github.com>